remove spring and cleanup pages
