+++ dev/package.json
@@ -37,7 +37,6 @@
     "copy-to-clipboard": "^3.3.3",
     "solid-devtools": "^0.24.7",
     "solid-js": "^1.6.8",
-    "solid-spring": "^0.0.7",
     "solid-start": "^0.2.11",
     "solid-start-trpc": "^0.0.16",
     "undici": "5.14.0",
